Run script
