+++ python/tank_vendor/yaml/resolver.py
